moved methods for connect and disconnect to basebleactivity
